+++ app/src/main/java/ca/josephroque/bowlingcompanion/App.kt
@@ -8,7 +8,6 @@ import android.arch.lifecycle.OnLifecycleEvent
 import android.content.Context
 import android.support.v7.preference.PreferenceManager
 import android.view.inputmethod.InputMethodManager
-import com.bugsnag.android.Bugsnag
 import com.google.android.gms.ads.MobileAds
 import java.util.concurrent.atomic.AtomicBoolean
 
@@ -47,10 +46,6 @@ class App : Application(), LifecycleObserver {
         super.onCreate()
         PreferenceManager.setDefaultValues(this, R.xml.pref_app, false)
         MobileAds.initialize(this, BuildConfig.ADMOB_APP_ID)
-
-        if (!BuildConfig.DEBUG) {
-            Bugsnag.init(this)
-        }
     }
 
     @OnLifecycleEvent(Lifecycle.Event.ON_STOP)